Extracted ResultSetResponseNormalizer
