+++ test/unit/models/classes/search/SearchProxyTest.php
@@ -22,14 +22,12 @@ declare(strict_types=1);
 
 namespace oat\tao\test\unit\model\search;
 
-use core_kernel_classes_Resource;
-use oat\generis\model\data\Ontology;
-use oat\generis\model\data\permission\PermissionHelper;
 use oat\generis\test\TestCase;
 use oat\tao\model\AdvancedSearch\AdvancedSearchChecker;
 use oat\tao\model\search\ElasticSearchBridge;
 use oat\tao\model\search\GenerisSearchBridge;
 use oat\tao\model\search\ResultSet;
+use oat\tao\model\search\ResultSetResponseNormalizer;
 use oat\tao\model\search\SearchProxy;
 use oat\tao\model\search\SearchQueryFactory;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -39,9 +37,6 @@ class SearchProxyTest extends TestCase
 {
     /** @var SearchProxy */
     private $subject;
-
-    /** @var PermissionHelper|MockObject */
-    private $permissionHelperMock;
 
     /** @var AdvancedSearchChecker|MockObject */
     private $advancedSearchCheckerMock;
@@ -61,82 +56,39 @@ class SearchProxyTest extends TestCase
     /** @var ServerRequestInterface|MockObject */
     private $requestMock;
 
-    /** @var Ontology|MockObject */
-    private $modelMock;
+    /** @var ResultSetResponseNormalizer|MockObject */
+    private $resultSetResponseNormalizerMock;
+
 
     public function setUp(): void
     {
-        $this->permissionHelperMock = $this->createMock(PermissionHelper::class);
         $this->advancedSearchCheckerMock = $this->createMock(AdvancedSearchChecker::class);
         $this->elasticSearchBridgeMock = $this->createMock(ElasticSearchBridge::class);
         $this->generisSearchBridgeMock = $this->createMock(GenerisSearchBridge::class);
         $this->searchQueryFactoryMock = $this->createMock(SearchQueryFactory::class);
+        $this->resultSetResponseNormalizerMock = $this->createMock(ResultSetResponseNormalizer::class);
 
         $this->resultSetMock = $this->createMock(ResultSet::class);
         $this->requestMock = $this->createMock(ServerRequestInterface::class);
-        $this->modelMock = $this->createMock(Ontology::class);
 
         $this->subject = new SearchProxy();
         $this->subject->setServiceLocator(
             $this->getServiceLocatorMock(
                 [
-                    PermissionHelper::class => $this->permissionHelperMock,
                     AdvancedSearchChecker::class => $this->advancedSearchCheckerMock,
                     ElasticSearchBridge::class => $this->elasticSearchBridgeMock,
                     GenerisSearchBridge::class => $this->generisSearchBridgeMock,
                     SearchQueryFactory::class => $this->searchQueryFactoryMock,
+                    ResultSetResponseNormalizer::class => $this->resultSetResponseNormalizerMock
                 ]
             )
         );
-
-        $this->subject->setModel($this->modelMock);
-
-        $this->resultSetMock
-            ->method('getArrayCopy')
-            ->willReturn(
-                [
-                    'uri1',
-                    'uri2',
-                ]
-            );
-
-        $this->permissionHelperMock
-            ->expects($this->once())
-            ->method('filterByPermission')
-            ->willReturn(
-                [
-                    'uri1',
-                ]
-            );
-
-        $resourceMock = $this->createMock(core_kernel_classes_Resource::class);
-
-        $resourceMock
-            ->method('getUri')
-            ->willReturn('uri1');
-
-        $resourceMock
-            ->method('getLabel')
-            ->willReturn('label');
-
-        $this->modelMock
-            ->method('getResource')
-            ->willReturn($resourceMock);
     }
 
     public function testSearchWithoutElasticSearch(): void
     {
-        $this->requestMock
-            ->method('getQueryParams')
-            ->willReturn(
-                [
-                    'rows' => 10,
-                    'page' => 1,
-                ]
-            );
-
         $this->advancedSearchCheckerMock
-            ->expects($this->exactly(2))
+            ->expects($this->once())
             ->method('isEnabled')
             ->willReturn(false);
 
@@ -145,14 +97,13 @@ class SearchProxyTest extends TestCase
             ->method('search')
             ->willReturn($this->resultSetMock);
 
-        $this->resultSetMock
-            ->expects($this->exactly(2))
-            ->method('getTotalCount')
-            ->willReturn(100);
+        $this->resultSetResponseNormalizerMock
+            ->expects($this->once())
+            ->method('normalize')
+            ->willReturn([]);
 
         $result = $this->subject->search($this->requestMock);
-        $this->assertResult($result);
-        $this->assertEquals(10.0, $result['total']);
+        $this->assertIsArray($result);
     }
 
     public function testSearchWithNoPagination(): void
@@ -162,7 +113,7 @@ class SearchProxyTest extends TestCase
             ->willReturn([]);
 
         $this->advancedSearchCheckerMock
-            ->expects($this->exactly(2))
+            ->expects($this->once())
             ->method('isEnabled')
             ->willReturn(true);
 
@@ -170,14 +121,13 @@ class SearchProxyTest extends TestCase
             ->method('search')
             ->willReturn($this->resultSetMock);
 
-        $this->resultSetMock
-            ->expects($this->exactly(1))
-            ->method('getTotalCount')
-            ->willReturn(100);
+        $this->resultSetResponseNormalizerMock
+            ->expects($this->once())
+            ->method('normalize')
+            ->willReturn([]);
 
         $result = $this->subject->search($this->requestMock);
-        $this->assertResult($result);
-        $this->assertEquals(1, $result['total']);
+        $this->assertIsArray($result);
     }
 
     public function testSearchWithElasticSearch(): void
@@ -192,7 +142,7 @@ class SearchProxyTest extends TestCase
             );
 
         $this->advancedSearchCheckerMock
-            ->expects($this->exactly(2))
+            ->expects($this->once())
             ->method('isEnabled')
             ->willReturn(true);
 
@@ -200,44 +150,12 @@ class SearchProxyTest extends TestCase
             ->method('search')
             ->willReturn($this->resultSetMock);
 
-        $this->resultSetMock
-            ->expects($this->exactly(2))
-            ->method('getTotalCount')
-            ->willReturn(100);
+        $this->resultSetResponseNormalizerMock
+            ->expects($this->once())
+            ->method('normalize')
+            ->willReturn([]);
 
         $result = $this->subject->search($this->requestMock);
-        $this->assertResult($result);
-        $this->assertEquals(10.0, $result['total']);
-    }
-
-    private function assertResult(array $result): void
-    {
         $this->assertIsArray($result);
-        $this->assertArrayHasKey('data', $result);
-        $this->assertArrayHasKey('readonly', $result);
-        $this->assertArrayHasKey('success', $result);
-        $this->assertArrayHasKey('page', $result);
-        $this->assertArrayHasKey('total', $result);
-        $this->assertArrayHasKey('totalCount', $result);
-        $this->assertArrayHasKey('records', $result);
-        $this->assertEquals(1, $result['page']);
-        $this->assertEquals(100, $result['totalCount']);
-        $this->assertEquals(2, $result['records']);
-        $this->assertCount(2, $result['data']);
-        $this->assertCount(1, $result['readonly']);
-        $this->assertEquals(
-            [
-                [
-                    'id' => 'uri1',
-                    'http://www.w3.org/2000/01/rdf-schema#label' => 'label',
-                ],
-                [
-                    'id' => 'uri1',
-                    'http://www.w3.org/2000/01/rdf-schema#label' => 'label',
-                ],
-            ],
-            $result['data']
-        );
-        $this->assertTrue($result['readonly']['uri2']);
     }
 }
